Support Bazel
Bazel is a language-agnostic build system. This patch adds a Dockerfile
that builds a solution using Bazel, as well as Dockerfile.base that
builds a base Docker image that contains various dependencies to avoid
accessing the Internet on building a solution.

While Bazel supports many programming languages, I verified only C++
builds for now. If other teams want to use Bazel with other languages
it is easy to add more distfiles.

Closes #12.